+++ src/client/render/gl/Renderer.ts
@@ -11,6 +11,7 @@
 
 import type { Config } from "../../../core/configuration/Config";
 import type { MapLayer } from "../../../core/game/TerrainMapLoader";
+import { translateText } from "../../Utils";
 import type { SpiralRibbon } from "../frame/SpiralTrails";
 import type {
   AttackRingInput,
@@ -640,6 +641,7 @@ export class GPURenderer {
       gl,
       header,
       this.paletteTex,
+      this.effectTex,
       this.settings,
     );
     this.structureLevelPass = new StructureLevelPass(gl, header, this.settings);
@@ -1182,6 +1184,12 @@ export class GPURenderer {
             cost: data.cost,
             canAfford: data.canAfford,
             canPlace: data.canBuild || data.canUpgrade,
+            topText:
+              data.multiplier && data.multiplier > 1
+                ? translateText("build_menu.upgrade_amount", {
+                    amount: data.multiplier.toString(),
+                  })
+                : undefined,
           }
         : null,
     );