+++ src/core/execution/ExecutionManager.ts
@@ -135,6 +135,7 @@ export class Executor {
           intent.unit,
           intent.tile,
           intent.rocketDirectionUp,
+          intent.amount,
           intent.troops,
         );
       case "allianceExtension": {
@@ -142,7 +143,11 @@ export class Executor {
       }
 
       case "upgrade_structure":
-        return new UpgradeStructureExecution(player, intent.unitId);
+        return new UpgradeStructureExecution(
+          player,
+          intent.unitId,
+          intent.amount,
+        );
       case "delete_unit":
         return new DeleteUnitExecution(player, intent.unitId);
       case "quick_chat":